requirements.txt
arabic-reshaper==3.0.0
attrs==22.2.0
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
decorator==5.1.1
einops==0.6.0
entrypoints==0.4
fastjsonschema==2.16.3
filelock==3.10.0
fire==0.5.0
fonttools==4.38.0
gast==0.5.3
h5py==3.8.0
huggingface-hub==0.13.3
idna==3.4
imageio==2.26.1
imgaug==0.4.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
lmdb==1.4.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
natsort==8.3.1
nbclient==0.7.2
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==23.0
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
PyArabic==0.6.15
pycuda==2022.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytools==2022.1.12
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.2
regex==2022.10.31
requests==2.28.2
safetensors==0.3.0
scikit-image==0.19.3
scipy==1.7.3
shapely==2.0.1
six==1.16.0
termcolor==2.2.0
tifffile==2021.11.2
timm==0.6.12
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
validators==0.20.0
Wand==0.6.11
zipp==3.15.0
